finishing up backend
