+++ models/playerUser.js
@@ -12,19 +12,8 @@ const userSchema = new Schema(
       unique: true,
     },
     name: { type: String, default: null, minlength: 2, maxlength: 128 },
-    education: {
-      type: String,
-      enum: userEducation,
-    },
     school: { type: String, default: null },
     password: { type: String, required: true },
-    country: { type: String, default: null },
-    region: { type: String, default: null },
-    city: { type: String, default: null },
-    chests_submitted: { type: Map, of: Number },
-    is_email_verified: { type: Boolean, default: false },
-    email_verification_token: { type: String, default: null },
-    reset_password_token: { type: String, default: null },
     role: {
       type: String,
       enum: userRoles,
